refactor pals viewmodel, layout file for requests page
