+++ app/src/main/java/com/example/packpals/viewmodels/PalsViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.packpals.views.pals
+package com.example.packpals.viewmodels
 
 import android.util.Log
 import androidx.lifecycle.LiveData
@@ -6,15 +6,13 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.example.packpals.models.Pal
 import com.example.packpals.models.PalRequest
-import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
 
-class FindAPalViewModel : ViewModel() {
+class PalsViewModel : ViewModel() {
     private val db = FirebaseFirestore.getInstance()
     private val usersCollectionRef = db.collection("pals")
-    private val auth = FirebaseAuth.getInstance()
 
-    val TAG = FindAPalViewModel::class.java.toString()
+    val TAG = PalsViewModel::class.java.toString()
 
     private val _palRequestsLiveData = MutableLiveData(listOf(
         PalRequest("0", "Tuff"),
@@ -35,6 +33,4 @@ class FindAPalViewModel : ViewModel() {
                 Log.w(TAG, "addPal error", exception)
             }
     }
-
-
 }
