+++ app/src/main/java/com/example/packpals/views/pals/FindAPalFragment.kt
@@ -7,6 +7,7 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.viewModels
 import com.example.packpals.R
+import com.example.packpals.viewmodels.PalsViewModel
 import kotlinx.android.synthetic.main.fragment_find_a_pal.view.findAPalSubmitButton
 import kotlinx.android.synthetic.main.fragment_find_a_pal.view.findAPalUsernameInput
 
@@ -16,7 +17,7 @@ import kotlinx.android.synthetic.main.fragment_find_a_pal.view.findAPalUsernameI
 class FindAPalFragment : Fragment() {
 
 //    private lateinit var recyclerView: RecyclerView
-    private val viewModel: FindAPalViewModel by viewModels()
+    private val viewModel: PalsViewModel by viewModels()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
